Merge depOrder.Aur and depOrder.Bases
Bases was a map, mapping the pkgbase name to all the packages we have
under that base. Aur was the order in which we installed the packages.
Bases was a map because it might have been usefull to query by name
instead of iterating over. Although this proved to never be needed, so
instead Bases can become an ordered slice, thus negating the need for
Aur.

The merge is under the name REALBASES, Aur and Bases will be removed and
replaced by this. The name REALBASES is used to that it is easy to sed
back to Aur once the former one has been removed.